app: generate git-version.h regardless of branch
Previously, git-version.h always recorded the latest commit on the
master branch. This change should record the current commit (which is
what gets built) regardless of the branch.

I've checked that it does "the right thing" regarding multi-parent
merges by checking out such a merge commit, generating git-version.h and
comparing it to the commit sha1 in the log.


Former-commit-id: 222ffd2360b80652ebffc44b5a00d1bfddf859d8