file-jpeg: Make cinfo struct static so asynchronous saves don't crash
The plug-in already makes a copy of the cinfo struct for asynchronous
save, but it seems that somewhere inside cinfo, a pointer points to
another region in this struct. So, everything is copied as-is, and
the pointer ends up pointing to somewhere in the save_image() stack
which is gone when that function returns.


Former-commit-id: b072e7583351171bae8297b8cd191ccd958bc6aa