configure: build with DISABLE_DEPRECATED also for GTK+ 2.24
Bump the "too new" test to GTK+ 2.26 which will never be released, but
this way we can keep the configure check around for later reuse after
the switch to GTK+ 3.x. Also bump the "too new" GLib and Pango
versions to really future versions so we early-catch deprecations
in the current development cycle.


Former-commit-id: 8d8c6bb9ea9b5e3a7df5c58a94c0304c6b08a0e4