app: Add a "check-for-deprecated-procedures-in-script-fu" make target
Add a "check-for-deprecated-procedures-in-script-fu" make target that
looks for usage of deprecated procedures in *.scm files. We currently
use quite a bit of deprecated procedures. When we don't do that
anymore, we should connect this to make check somehow so our nightly
build fails if someone deprecated a procedures without also porting
all clients.

Run it like this from the source root:

  make -C plug-ins/script-fu check-for-deprecated-procedures-in-script-fu


Former-commit-id: 28983a5158196eb634050f8e22439cc7fc640b6a