app: always add ALT+number accelerators to the image window menu items
There is no reason why we should have them only in single-window-mode.
The accels are numbered in the order the displays are opened, which is
also the order they appear in the menu. The only thing that needs to
be taken care of is that switching to SWM swallows the image windows
into the single window in the right order. Fixes bug #647524.


Former-commit-id: 3b71b71c4cddc561e2b258a5226a2dfa6b95a6bf