file-jpeg: (save) Change default coding mode to progressive
This was discussed in #gimp. This would no longer result in strictly
baseline JPEGs, but progressive mode has been supported for about
a decade in browsers and libjpeg now.

Progressive mode brings two advantages: (1) Medium to large images get
decent space savings to due to coding of large sequences of
zero coefficients, and (2) The images load progressively in a browser
which leads to better user experience when viewing medium to large
images on a website.

Those who want strictly baseline have the option to turn it off.


Former-commit-id: 4b3a1c6680ba6890b5b35fb8ba6e3e46e7257097