devel-docs: Add gegl-porting-plan.txt
In order to effectively collaborate on this huge porting task, the
first step is to agree on exactly what we want to do.

In an effort to organize this, add gegl-porting-plan.txt to
devel-docs. In this document, we can describe how we want the port to
be done. This is not only for ourselves, but will also be useful for
non-core contributors.


Former-commit-id: 0f0bb336a86946cb22861e265d1a7def0af45b71