file-compressor: Add support for xz compressed files (bug #683159)
This commit causes a compiler warning because string literals are of
type signed char, and we use a character out of the signed range in the
mimetype for xz.  It can be fixed by changing the data type of the mime
string to (unsigned char *), and not using string literals, but that's
a bit too much for this commit.


Former-commit-id: fc4a192f23bb44a59e199118c8c8a7d3027d6fd1