plug-ins: give animation playback ability to take into account image updates.
This includes a much better code organization where initialization part can be reused
to free up memory and refresh the display on the fly.
This refresh feature has a UI button and a shortcut (ctrl-r) associated.
It implies any kind of change, additional or removed layers, new image size, etc.


Former-commit-id: 2bfae573adab409d59f3b4d354f31c185f94ba21