file-gif-load: Fix incorrect out-of-input-bits condition (#695513)
GetCode() erroneously assumes that it has run out of input bits, because
it's off by 1 in its size calculation. At the end of the block, if the
End-of-Information code is exactly at the block boundary, it calculates
that it needs to read an extra bit, and tries to read beyond the end of
the block. It fails then, and the End-of-Information code is not
processed, and other warnings/errors are generated.


Former-commit-id: f86655bb1fcf356524c1894096f4866633cc7666