app: experimental: use the new coordinate picking in property tables
If we add a chain button to link x/y or width/height properties, also
add a coordinate pick button for these properties. This looks totally
ugly and is broken for most cases where it happens, but it works
nicely for e.g. gegl:fractal-explorer or gegl:grid. This clearly needs
some metadata for operation properties to be useful.


Former-commit-id: 38e9c79d1e1ab725939cf12becd435c37e1bff74