Bug 576910: Show 'Language' setting languages in native language
The trick works by temporarily resetting the current locale to localize
each language label in its own lang.
One exception is English that is equivalent to the "C" code, and we make
also some special exception for Chinese where there are very different
variant depending on the region.
I also ensure the "System Language" string is translated in whatever
language is the system actually set to.


Former-commit-id: f6dcde1ee66fda4dfdc063022c4d2e901adb9a71